refactor: remove system-tests crate
