--- system-tests/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "system-tests"
-publish = false
-edition.workspace = true
-version.workspace = true
-license.workspace = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-clap.workspace = true
-unimarkup-core = { path = "../core/" }
-shlex = "1.1.0"
