--- system-tests/README.md
@@ -1,3 +1,0 @@
-# System tests
-
-This section contains system tests for the Unimarkup reference implementation.
